backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
psutil==7.2.1
scapy==2.7.0
requests==2.32.5
python-dotenv==1.2.1
openai==2.15.0
pydantic==2.12.5
httpx==0.28.1
psycopg[binary]==3.3.2
pysnmp==7.1.22